feat: Make final score no longer be rounded by 10 (#287)
## Description

This PR makes the time bonus more precise, so scores are no longer
limited to multiples of 10. The time bonus now keeps its fractional
value until the final score is rounded.

Closes #173.

Supersedes #263.

+++ CutTheRopeDX/GameMain/BoxOpenClose.cs
@@ -725,7 +725,7 @@ namespace CutTheRopeDX.GameMain
         public int raState;
 
         /// <summary>Time bonus score used by the result countdown.</summary>
-        public int timeBonus;
+        public float timeBonus;
 
         /// <summary>Star bonus score used by the result countdown.</summary>
         public int starBonus;
